Refactored code and completed documentation
