+++ .changeset/chilled-teachers-bathe.md
@@ -1,0 +1,5 @@
+---
+'@as-integrations/aws-lambda': major
+---
+
+Implement strictly type request handlers, custom handlers, and expose middleware functionality
